Move AdvancedTab to gui package
