+++ src/main/java/net/sf/jabref/PrefsDialog3.java
@@ -39,6 +39,7 @@ import javax.swing.event.ListSelectionListener;
 
 import net.sf.jabref.export.ExportFormats;
 import net.sf.jabref.groups.GroupsPrefsTab;
+import net.sf.jabref.gui.AdvancedTab;
 import net.sf.jabref.gui.FileDialogs;
 import net.sf.jabref.gui.MainTable;
